Move `PropertyTypeStore` to `store` crate
