+++ libs/@local/graph/store/src/property_type/mod.rs
@@ -1,7 +1,16 @@
 pub(crate) use self::query::PropertyTypeQueryPathVisitor;
-pub use self::query::{PropertyTypeQueryPath, PropertyTypeQueryToken};
+pub use self::{
+    query::{PropertyTypeQueryPath, PropertyTypeQueryToken},
+    store::{
+        ArchivePropertyTypeParams, CountPropertyTypesParams, CreatePropertyTypeParams,
+        GetPropertyTypeSubgraphParams, GetPropertyTypeSubgraphResponse, GetPropertyTypesParams,
+        GetPropertyTypesResponse, PropertyTypeStore, UnarchivePropertyTypeParams,
+        UpdatePropertyTypeEmbeddingParams, UpdatePropertyTypesParams,
+    },
+};
 
 mod query;
+mod store;
 
 use hash_graph_types::ontology::PropertyTypeWithMetadata;
